Changed Logging deque to receiver.
